+++ test/fixtures/vue-runtime-entrypoint/index.js
@@ -1,0 +1,7 @@
+import Vue from 'vue';
+import createFilePond from 'vue-filepond';
+
+export {
+    Vue,
+    createFilePond
+};